fix character config
