+++ src/App.tsx
@@ -10,7 +10,7 @@ const Body = lazy(() => import('./pages/Home'))
 const TeamPage = lazy(() => import('./pages/TeamDetails'))
 const Rotation = lazy(() => import('./Rotation'))
 // const Test = lazy(() => import('./TestPage'))
-const CharacterConfig = lazy(() => import('./components/CharacterConfig'))
+const CharacterConfig = lazy(() => import('./pages/CharacterConfig'))
 
 function Content() {
     const dbContextObj = new DatabaseContext(new DBLocalStorage(localStorage));
@@ -26,10 +26,14 @@ function Content() {
                         <BrowserRouter basename="/">
                             <Routes>
                                 <Route index element={<Body />} />
-                                <Route path="/TeamPage/:teamIndex" element={<TeamPage />} />
+                                <Route path="/TeamPage/*">
+                                    <Route path=":teamIndex/*">
+                                        <Route index element={<TeamPage />} />
+                                        <Route path=":charIndex" element={<CharacterConfig />} />
+                                    </Route>
+                                </Route>
                                 <Route path="/rotationPage" element={<Rotation />} />
                                 {/* <Route path="/testPage" element={<Test />} /> */}
-                                <Route path="/characterConfig" element={<CharacterConfig />} />
                             </Routes>
                         </BrowserRouter>
                     </Suspense>
